feat: moved commands from being initialized from init functions to classic create functions
