--- pkg/commands/app/app_commands.go
@@ -1,10 +1,0 @@
-package app
-
-import (
-	"github.com/qlik-oss/corectl/cmd"
-	"github.com/spf13/cobra"
-)
-
-func EditAppSubCommand() *cobra.Command {
-	return cmd.EditAppSubCommand()
-}
