+++ pkg/commands/engine/flags.go
@@ -1,4 +1,4 @@
-package cmd
+package engine
 
 import (
 	"fmt"
@@ -11,6 +11,9 @@ import (
 
 var localFlags pflag.FlagSet
 var initialized bool
+
+// DefaultUnbuildFolder is the placeholder for unbuild folder location
+var DefaultUnbuildFolder = "./<app name>-unbuild"
 
 func withLocalFlags(ccmd *cobra.Command, localFlagNames ...string) *cobra.Command {
 	if !initialized {
@@ -54,6 +57,7 @@ func initLocalFlags() {
 	localFlags.String("objects", "", "A list of generic object json paths")
 	localFlags.String("script", "", "Path to a qvs file containing the app data reload script")
 	localFlags.String("app-properties", "", "Path to a json file containing the app properties")
+	localFlags.String("dir", DefaultUnbuildFolder, "Path to a the folder where the unbuilt app is exported")
 
 	if runtime.GOOS != "windows" {
 		// Set annotation to run bash completion function
